Use url path instead of query param
Currently the fragment name is passed via query param. e.g.  `header`
would be `?fragment=header`.

This removes the query param in favor of extending the target url,
creating a unique url for each fragment.

e.g. the `site/header` fragment with a target of
`https://localhost:3000/_view_fragments` would result in a request to
`htttps://localhost:3000/_view_fragments/site/header`

This aligns the behavior of fragments with the behavior of layouts.