Add support for HMAC request verification
This adds support for sending an `Authorization` header to the target
server when requesting page fragments. This should help prevent
unauthorized access of fragment/layout endpoints.

A timestamp is included as an `X-Authorization-Time` header which is
also part of the HMAC itself, allowing target servers to implement their
own expiration logic.