Add support for gzipped responses
When forwarding headers the requesting browser will sometimes request,
and receive a gzipped response from the server. viewproxy does not
currently look for or handle gzipped responses which leads to strange
behavior. e.g. only the layout rendering.

To fix this, when reading the response body from a fragment or layout
request we now check the `Content-Encoding` header to see if the request
is gzipped. If the request is gzipped, we gunzip the response body as
part of the fragment building process.

Additionally, if the layout request was gzipped then the response we
send to the user will also be gzipped.