Add multiplexer.Tripper interface
This adds a transport interface that allows consumers of viewproxy to
provide their own logic for fetching fragments.

Instead of providing completely paved paths with logging and
instrumentation this will allow consumers to build their own low-level
logging, instrumentation, http request logic, and more.

A default tripper implementation is provided that delegates to an
`http.Client{}`.

Additionally, the `route`, and `fragment` for the request is provided in
the `request.Context()` so consumers can log or instrument individual
requests.