Add support for dynamic fragment path replacement
The current implementation of fragments assumes that all fragments will
be static paths with no dynamic parts. Dynamic parts are instead passed
as query parameters to the static path provided.

This changes fragments to support dynamic parts. When constructing a
requestable from a fragment a target `url.URL` and a map of named path
parameters can be provided to generate a valid URL to request in the
multiplexer.

This does not yet support query parameters.