Add configuration for ignoring trailing slashes
This adds a configuration option to `viewproxy.Server` that allows you
to ignore the trailing slash on incoming requests.

Given the following route:

```
viewproxy.Get("/foo/:bar")
```

The following two routes will be considered the same:

`http://mysite.net/foo/bar`
`http://mysite.net/foo/bar/`

A route can partially opt-out of this behavior by being declared with a
trailing `/`. In that case, a trailing slash will be required for the
route to match.